Almost done making lock file configurable in loaders

+++ loaders/loader.go
@@ -149,8 +149,9 @@ func (ldr *Loader) ProcessFolder(ctx *context.Context, caching, loadOption strin
 
 // lockFolder will attempt to lock the folder by creating the lock file
 func (ldr *Loader) lockFolder() (err error) {
-	if _, err = os.Stat(ldr.lockFilepath); err != nil && os.IsNotExist(err) {
-		return fmt.Errorf("file: %v not found", ldr.lockFilepath)
+	// If the path is an empty string, we should not be locking
+	if ldr.lockFilepath == utils.EmptyString {
+		return
 	}
 	_, err = os.OpenFile(ldr.lockFilepath,
 		os.O_RDONLY|os.O_CREATE, 0644)
@@ -158,6 +159,11 @@ func (ldr *Loader) lockFolder() (err error) {
 }
 
 func (ldr *Loader) unlockFolder() (err error) {
+	// If the path is an empty string, we should not be locking
+	if ldr.lockFilepath == utils.EmptyString {
+		return
+	}
+
 	if _, err = os.Stat(ldr.lockFilepath); err == nil {
 		return os.Remove(ldr.lockFilepath)
 	}
@@ -165,6 +171,10 @@ func (ldr *Loader) unlockFolder() (err error) {
 }
 
 func (ldr *Loader) isFolderLocked() (locked bool, err error) {
+	// If the path is an empty string, we should not be locking
+	if ldr.lockFilepath == utils.EmptyString {
+		return
+	}
 	if _, err = os.Stat(ldr.lockFilepath); err == nil {
 		return true, nil
 	}
